+++ services/session/ipc/stub/avsession_controller_stub.cpp
@@ -131,7 +131,7 @@ int32_t AVSessionControllerStub::HandleGetAVMetaData(MessageParcel& data, Messag
     std::shared_ptr<AVSessionPixelMap> mediaPixelMap = metaData.GetMediaImage();
     if (mediaPixelMap != nullptr) {
         mediaImageBuffer = mediaPixelMap->GetInnerImgBuffer();
-        mediaImageLength = mediaImageBuffer.size();
+        mediaImageLength = static_cast<int>(mediaImageBuffer.size());
         metaData.SetMediaLength(mediaImageLength);
     }
 
@@ -140,7 +140,7 @@ int32_t AVSessionControllerStub::HandleGetAVMetaData(MessageParcel& data, Messag
     std::shared_ptr<AVSessionPixelMap> avQueuePixelMap = metaData.GetAVQueueImage();
     if (avQueuePixelMap != nullptr) {
         avQueueImageBuffer = avQueuePixelMap->GetInnerImgBuffer();
-        avQueueImageLength = avQueueImageBuffer.size();
+        avQueueImageLength = static_cast<int>(avQueueImageBuffer.size());
         metaData.SetAVQueueLength(avQueueImageLength);
     }
